refac(crypto): extract `PoseidonSpongeBase` from `poseidon.h`
